backend/requirements.txt
fastapi
uvicorn
pytesseract
Pillow
python-dotenv
google-generativeai
